refactor: fold fragment extension files into their base files
Eleven satellite files were splinters of the old SwiftLint
type_body_length rule (since removed) rather than cohesive topics —
several held 2-3 trivial methods:

- GlobalSettings: +Volume, +SoloMode, +PlatformSettings, +AudioSession,
  +Logging -> GlobalSettings.swift (302 lines)
- AudioManager: +Initialization, +Persistence, +SoundManagement ->
  AudioManager.swift (281 lines)
- SoundSheetForm+Helpers -> SoundSheetForm.swift
- MixerView+Layouts -> MixerView.swift
- SoundsListTemplate+ImageRendering -> SoundsListTemplate.swift

Code moves verbatim as same-file extension blocks under MARKs; only
satellite headers and duplicate imports were dropped. Large
single-topic extension files (SoloMode, MediaControls, etc.) and
framework extensions are untouched.
